+++ src/Program.cs
@@ -2,23 +2,12 @@
 // Licensed under the MIT License.
 
 using System.CommandLine.Builder;
+using AzureMcp.Areas;
 using AzureMcp.Commands;
-using AzureMcp.Services.Azure.AppConfig;
-using AzureMcp.Services.Azure.Cosmos;
-using AzureMcp.Services.Azure.KeyVault;
-using AzureMcp.Services.Azure.Kusto;
-using AzureMcp.Services.Azure.Monitor;
-using AzureMcp.Services.Azure.Postgres;
-using AzureMcp.Services.Azure.Redis;
-using AzureMcp.Services.Azure.Aks;
 using AzureMcp.Services.Azure.ResourceGroup;
-using AzureMcp.Services.Azure.Search;
-using AzureMcp.Services.Azure.ServiceBus;
-using AzureMcp.Services.Azure.Storage;
 using AzureMcp.Services.Azure.Subscription;
 using AzureMcp.Services.Azure.Tenant;
 using AzureMcp.Services.Caching;
-using AzureMcp.Services.Interfaces;
 using AzureMcp.Services.ProcessExecution;
 using AzureMcp.Services.ToolInstallation;
 using Microsoft.Extensions.DependencyInjection;
@@ -26,6 +15,8 @@ using Microsoft.Extensions.Logging;
 
 internal class Program
 {
+    private static IAreaSetup[] Areas = RegisterAreas();
+
     private static async Task<int> Main(string[] args)
     {
         try
@@ -54,6 +45,36 @@ internal class Program
             });
             return 1;
         }
+    }
+    private static IAreaSetup[] RegisterAreas()
+    {
+
+        return [
+            // Register core areas
+            new AzureMcp.Areas.AzureBestPractices.AzureBestPracticesSetup(),
+            new AzureMcp.Areas.Extension.ExtensionSetup(),
+            new AzureMcp.Areas.Group.GroupSetup(),
+            new AzureMcp.Areas.Server.ServerSetup(),
+            new AzureMcp.Areas.Subscription.SubscriptionSetup(),
+            new AzureMcp.Areas.Tools.ToolsSetup(),
+
+            // Register Azure service areas
+            new AzureMcp.Areas.AppConfig.AppConfigSetup(),
+            new AzureMcp.Areas.Authorization.AuthorizationSetup(),
+            new AzureMcp.Areas.AzureIsv.AzureIsvSetup(),
+            new AzureMcp.Areas.Cosmos.CosmosSetup(),
+            new AzureMcp.Areas.Foundry.FoundrySetup(),
+            new AzureMcp.Areas.KeyVault.KeyVaultSetup(),
+            new AzureMcp.Areas.Kusto.KustoSetup(),
+            new AzureMcp.Areas.Monitor.MonitorSetup(),
+            new AzureMcp.Areas.Postgres.PostgresSetup(),
+            new AzureMcp.Areas.Redis.RedisSetup(),
+            new AzureMcp.Areas.Search.SearchSetup(),
+            new AzureMcp.Areas.ServiceBus.ServiceBusSetup(),
+            new AzureMcp.Areas.Sql.SqlSetup(),
+            new AzureMcp.Areas.Storage.StorageSetup(),
+            new AzureMcp.Areas.BicepSchema.BicepSchemaSetup(),
+        ];
     }
 
     private static Parser BuildCommandLineParser(IServiceProvider serviceProvider)
@@ -86,28 +107,23 @@ internal class Program
     private static void WriteResponse(CommandResponse response)
     {
         Console.WriteLine(JsonSerializer.Serialize(response, ModelsJsonContext.Default.CommandResponse));
-    }    internal static void ConfigureServices(IServiceCollection services)
+    }
+    internal static void ConfigureServices(IServiceCollection services)
     {
         services.ConfigureOpenTelemetry();
+
         services.AddMemoryCache();
         services.AddSingleton<ICacheService, CacheService>();
         services.AddSingleton<IExternalProcessService, ExternalProcessService>();
-        services.AddSingleton<IToolInstallationService, ToolInstallationService>();
-        services.AddSingleton<ISubscriptionService, SubscriptionService>();
         services.AddSingleton<ITenantService, TenantService>();
-        services.AddSingleton<ICosmosService, CosmosService>();
-        services.AddSingleton<IKustoService, KustoService>();
-        services.AddSingleton<IStorageService, StorageService>();
-        services.AddSingleton<IMonitorService, MonitorService>();
-        services.AddSingleton<IMonitorHealthModelService, MonitorHealthModelService>();
         services.AddSingleton<IResourceGroupService, ResourceGroupService>();
-        services.AddSingleton<IAppConfigService, AppConfigService>();
-        services.AddSingleton<ISearchService, SearchService>();
-        services.AddSingleton<IPostgresService, PostgresService>();
-        services.AddSingleton<IKeyVaultService, KeyVaultService>();
-        services.AddSingleton<IServiceBusService, ServiceBusService>();
-        services.AddSingleton<IRedisService, RedisService>();
-        services.AddSingleton<IAksService, AksService>();
+        services.AddSingleton<ISubscriptionService, SubscriptionService>();
         services.AddSingleton<CommandFactory>();
+
+        foreach (var area in Areas)
+        {
+            services.AddSingleton(area);
+            area.ConfigureServices(services);
+        }
     }
 }